requirements.txt
Django==4.0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
idna==2.9
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2022.5
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
Pillow==9.2.0
PyYAML>=5.3.1
django-rest-swagger==2.2.0
django-dotenv==1.4.2

#opencv
opencv-python
numpy==1.24.2
cvzone==1.5.6
ultralytics==8.0.26
hydra-core>=1.2.0
matplotlib>=3.2.2
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
filterpy==1.4.5
scikit-image==0.19.3
lap==0.4.0
